rename files to correctly reflect structs
