+++ calyx-frontend/src/lib.rs
@@ -5,12 +5,12 @@
 //! in the `ir` module.
 
 pub mod ast;
-mod attribute;
+mod attributes;
 mod common;
 pub mod parser;
 mod workspace;
 
 pub use ast::NamespaceDef;
-pub use attribute::{Attributes, GetAttributes};
+pub use attributes::{Attributes, GetAttributes};
 pub use common::{Direction, PortDef, Primitive, Width};
 pub use workspace::Workspace;
